server/requirements.txt
Flask==0.10.1
livereload==2.4.0
mysql-python==1.2.5
flask-mysql==1.3
flask-sqlalchemy==2.0
requests==2.7.0
lxml==3.4.4
alembic==0.8.2